+++ streamdeck_ui/display/filter.py
@@ -36,7 +36,7 @@ class Filter(ABC):
     @abstractmethod
     def transform(self, get_input: Callable[[], Image.Image], get_output: Callable[[int], Image.Image], input_changed: bool, time: Fraction) -> Tuple[Image.Image, int]:
         """
-        Transforms the given input image to te desired output image.
+        Transforms the given input image to the desired output image.
         The default behaviour is to return the orignal image.
 
         :param Callable[[], PIL.Image] get_input: A function that returns the input image to transform. Note that calling