+++ streamdeck_ui/stream_deck_monitor.py
@@ -74,10 +74,8 @@ class StreamDeckMonitor:
         self.streamdecks = {}
 
     def _run(self):
-        """Runs the internal monitor thread until completion
-        """
+        """Runs the internal monitor thread until completion"""
         while not self.quit.is_set():
-
             with self.lock:
                 attached_streamdecks = DeviceManager.DeviceManager().enumerate()