+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Change log
 ### Fixes
 - Requirement for Python < 3.11 removed.
 - Switched to pyside6.
+- Better support for dark/light mode (due to pyside6).
 ### Features
 - Added support for a new sub-variant of the StreamDeck XL.
 ## 2.0.6 - 23 Sep 2022